pyproject.toml
[project]
name = "brmed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
